+++ frontend/.env
@@ -1,0 +1,1 @@
+REACT_APP_GOOGLE_MAPS_API_KEY = AIzaSyBd9VunbiAT2b0r3FjRj-djyOVPf_zhp6Q
