+++ frontend/src/components/TripList/NewTripModal/DestinationList/index.jsx
