+++ frontend/src/components/TripList/NewTripModal/Searchbar/index.jsx
