+++ frontend/src/components/TripList/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Trip from "./Trip";
+import Trip from "./NewTripModal";
 
 function TripList() {
     return (
